backend/pyproject.toml
[project]
name = "lablocal"
version = "0.1.0"
description = "Visor de analíticas médicas autoalojado — self-hosted, open source (AGPL-3.0-or-later)"
license = {text = "AGPL-3.0-or-later"}
requires-python = ">=3.12"
dependencies = [
    "Django>=5.2,<6.0",
    "whitenoise>=6.7",
    "gunicorn>=22.0",
    "python-dotenv>=1.0",
    "dj-database-url>=2.1",
    "pillow>=12.2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.8",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "medivault.settings"
python_files = ["tests.py", "test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"
